requirements.txt
Flask
Flask-Mail
Flask-SQLAlchemy
itsdangerous
Werkzeug
python-dotenv
Flask-Migrate
